backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
transformers
torch
pydantic>=2.0
python-multipart

# Google Calendar integration
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv

# OpenAI chat integration
openai>=1.3.0
